requirements.txt
aspose-zip==24.2.0